refactor: split controllers and services to prevent model and pojo namespace collisions
